Started implementing an AbstractEnemy class, this class should be able to be extended and implemented on all enemies.
It provides properties to store Enemy MoveSpeed, Attack, Health, CurrentHealth, and EnemyName.

I am working on this still. I need to take a deeper look at it, and double check the collision still. Last I checked it was broken, but this should allow us to more easily implement enemies in the future.